renamed kuzzleSearchResult.js -> SearchResult.js
