+++ src/SearchResult.js
